requirements.txt
boto3
argparse
configparser
pyyaml
paramiko
